+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 This file only records notable changes. Not synchronized with all releases and tags.
 
 - v4 - not yet released
+  - Change default delay and intervals: [#1148](https://github.com/kachick/wait-other-jobs/pull/1148)
   - Drop deprecated options
   - Update action engine to Node.js v24
 - v3.8.0