+++ CHANGELOG.md
@@ -2,12 +2,14 @@
 
 This file only records notable changes. Not synchronized with all releases and tags.
 
-- v4 - not yet released
+- v4 - not yet released # See also [migration guide](docs/migration-v4.md)
   - Change default delay and intervals: [#1148](https://github.com/kachick/wait-other-jobs/pull/1148)
   - Drop to support key-value format in startupGracePeriod: [GH-1149](https://github.com/kachick/wait-other-jobs/pull/1149)
   - Support `event-list` to wait only for same event jobs by default (e.g. `push` -> `push`): [GH-998](https://github.com/kachick/wait-other-jobs/issues/998)
-  - Drop deprecated options
+  - Drop deprecated options: `wait-seconds-before-first-polling` and `min-interval-seconds`
+  - Deprecate `eventName` in favor of `eventNames` in `wait-list` and `skip-list`
   - Update action engine to Node.js v24
+  - Allow comments and trailing commas in JSON inputs
 - v3.8.0
   - Deprecate some options to prepare v4: [#1045](https://github.com/kachick/wait-other-jobs/pull/1045)
   - Add `outputs.parameters`