+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 This file only records notable changes. Not synchronized with all releases and tags.
 
-- v4 - not yet released # See also [migration guide](docs/migration-v4.md)
+- v4.0.0 # See also [migration guide](docs/migration-v4.md)
   - Change default delay and intervals: [#1148](https://github.com/kachick/wait-other-jobs/pull/1148)
   - Drop to support key-value format in startupGracePeriod: [GH-1149](https://github.com/kachick/wait-other-jobs/pull/1149)
   - Support `event-list` to wait only for same event jobs by default (e.g. `push` -> `push`): [GH-998](https://github.com/kachick/wait-other-jobs/issues/998)